+++ CHANGELOG.md
@@ -51,6 +51,8 @@
         - Increase minimum password length to eight.
     - Changes
         - Send contact form emails from do-not-reply address if sender's domain uses DMARC.
+    - New features:
+        - Roles can now have category restrictions like users.
 
 * v3.1 (16th November 2020)
     - Security: